backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pyzbar==0.1.9
opencv-python==4.9.0.80
mysql-connector-python==8.4.0
python-dotenv==1.0.1
sqlalchemy>=1.4.0,<2.0

